--- test/step8.c
@@ -1,76 +1,0 @@
-#include <stdio.h>
-#include <stddef.h>
-#include <signal.h>
-#include <unistd.h>
-
-#include "util.h"
-#include "net.h"
-#include "ip.h"
-#include "ip6.h"
-
-#include "driver/loopback.h"
-
-#include "test.h"
-
-static volatile sig_atomic_t terminate;
-
-static void
-on_signal(int s)
-{
-    (void)s;
-    terminate = 1;
-}
-
-static void
-cleanup(void)
-{
-    net_shutdown();
-}
-
-int
-main(int argc, char *argv[])
-{
-    ip6_addr_t src, dst;
-    size_t offset = IPV6_HDR_SIZE;
-
-    struct net_device *dev;
-    struct ip6_iface *iface;
-
-    signal(SIGINT, on_signal);
-    if (net_init() == -1) {
-        errorf("net_init() failure");
-        return -1;
-    }
-    dev = loopback_init();
-    if (!dev) {
-        errorf("loopback_init() failure");
-        return -1;
-    }
-    iface = ip6_iface_alloc(LOOPBACK_IPV6_ADDR, LOOPBACK_IPV6_NETMASK);
-    if (!iface) {
-        errorf("ip6_iface_alloc() failure");
-        return -1;
-    }
-    if (ip6_iface_register(dev, iface) == -1) {
-        errorf("ip6_iface_register() failure");
-        return -1;
-    }
-    if (net_run() == -1) {
-        errorf("net_run() failure");
-        return -1;
-    }
-
-    ip6_addr_pton(LOOPBACK_IPV6_ADDR, &src);
-    ip6_addr_pton(LOOPBACK_IPV6_ADDR, &dst);
-    while (!terminate) {
-        //if (ip6_output(IP_PROTOCOL_ICMPV6, test_data, sizeof(test_data), src, dst) == -1) {
-        if (ip6_output(IPV6_NEXT_ICMPV6, test_data + offset, sizeof(test_data) - offset, src, dst) == -1) {
-        //if (ip6_output(IP_PROTOCOL_ICMPV6, NULL, 40, src, dst) == -1) {
-            errorf("failure");
-            break;
-        }
-        sleep(1);
-    }
-    cleanup();
-    return 0;
-}
