+++ test/test.h
@@ -3,40 +3,28 @@
 
 #include <stdint.h>
 
+/* loopback */
 #define LOOPBACK_IP_ADDR "127.0.0.1"
 #define LOOPBACK_NETMASK "255.0.0.0"
-//#define LOOPBACK_IP_NETMASK "255.0.0.0"
-
 #define LOOPBACK_IPV6_ADDR "::1"
 #define LOOPBACK_IPV6_NETMASK "ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff"
 
+/* tap */
 #define ETHER_TAP_NAME    "tap0"
 #define ETHER_TAP_HW_ADDR "00:00:5e:00:53:01"
 #define ETHER_TAP_IP_ADDR "192.0.2.2"
 #define ETHER_TAP_NETMASK "255.255.255.0"
-//#define ETHER_TAP_IPV4_NETMASK "255.255.255.0"
-#define ETHER_TAP_IPV6_ADDR "fd00:cafe::2"
+#define ETHER_TAP_IPV6_ADDR "2001:db8::2"
 #define ETHER_TAP_IPV6_NETMASK "ffff:ffff:ffff:ffff::"
 
-#define DEFAULT_GATEWAY "192.0.2.1"
-#define DEFAULT_GATEWAY6 "fd00:cafe::1"
+/* raw */
+#define ETHER_PCAP_NAME    "Host1-veth0"
+#define ETHER_PCAP_HW_ADDR "00:00:5e:00:53:01"
+#define ETHER_PCAP_IPV6_ADDR "fd00:cafe:dead:1::2"
+#define ETHER_PCAP_IPV6_NETMASK "ffff:ffff:ffff:ffff::"
 
-/*
-const uint8_t test_data[] = {
-    0x45, 0x00, 0x00, 0x30,
-    0x00, 0x80, 0x00, 0x00,
-    0xff, 0x01, 0xbd, 0x4a,
-    0x7f, 0x00, 0x00, 0x01,
-    0x7f, 0x00, 0x00, 0x01,
-    0x08, 0x00, 0x35, 0x64,
-    0x00, 0x80, 0x00, 0x01,
-    0x31, 0x32, 0x33, 0x34,
-    0x35, 0x36, 0x37, 0x38,
-    0x39, 0x30, 0x21, 0x40,
-    0x23, 0x24, 0x25, 0x5e,
-    0x26, 0x2a, 0x28, 0x29
-};
-*/
+#define DEFAULT_GATEWAY "192.0.2.1"
+#define IPV6_DEFAULT_GATEWAY "fd00:cafe:dead:1::1"
 
 const uint8_t test_data[] = {
     0x60, 0x0a, 0x09, 0x00,
